+++ noir-r1cs/src/compiler.rs
@@ -1,21 +1,10 @@
 use {
-    crate::{sparse_matrix::mat_mul, SparseMatrix}, acir::{
+    crate::{r1cs_matrices::R1CSMatrices, solver::{R1CSSolver, WitnessBuilder}}, acir::{
         circuit::{opcodes::BlockType, Circuit, Opcode},
-        native_types::{Expression, Witness as AcirWitness, WitnessMap},
+        native_types::{Expression, Witness as AcirWitness},
         AcirField, FieldElement,
-    }, rand::{seq::index, Rng}, serde::{Deserialize, Serialize}, std::{collections::BTreeMap, fmt::{Debug, Formatter}, fs::File, io::Write, ops::Neg, vec}
+    }, std::{collections::BTreeMap, fmt::{Debug, Formatter}, ops::Neg, vec}
 };
-
-#[derive(Serialize)]
-struct JsonR1CS {
-    num_public:      usize,
-    num_variables:   usize,
-    num_constraints: usize,
-    a:               Vec<MatrixEntry>,
-    b:               Vec<MatrixEntry>,
-    c:               Vec<MatrixEntry>,
-    witnesses:       Vec<Vec<String>>,
-}
 
 #[derive(Debug, Clone)]
 /// Used for tracking reads of a read-only memory block
@@ -26,33 +15,6 @@ pub struct ReadOnlyMemoryBlock {
     pub static_reads: Vec<(usize, usize)>,
     /// (R1CS witness index of address, R1CS witness index of value read) tuples
     pub dynamic_reads: Vec<(usize, usize)>,
-}
-
-#[derive(Debug, Clone)]
-/// Indicates how to solve for an R1CS witness value in terms of earlier R1CS witness values and/or
-/// ACIR witness values.
-pub enum WitnessBuilder {
-    /// Constant value, used for the constant one witness & e.g. static lookups
-    Constant(FieldElement),
-    /// A witness value carried over from the ACIR circuit
-    Acir(usize),
-    /// A Fiat-Shamir challenge value
-    Challenge,
-    /// The inverse of the value at a specified witness index
-    Inverse(usize),
-    /// The sum of many witness values
-    Sum(Vec<usize>),
-    /// The product of the values at two specified witness indices
-    Product(usize, usize),
-    /// Witness is the result of a memory read from the .0th block at the .1th static address, whose value is available as the .2th acir witness index
-    StaticMemoryRead(usize, usize, usize),
-    /// Witness is the result of a memory read from the .0th block at the address determined by the .1th R1CS witness, whose value is available as the .2th acir witness index
-    DynamicMemoryRead(usize, usize, usize),
-    /// The number of times that the .1th index of the .0th memory block is accessed
-    MemoryAccessCount(usize, usize),
-    /// For solving for the denominator of an indexed lookup.
-    /// Fields are (sz_challenge, (index_coeff, index), rs_challenge, value).
-    LogUpDenominator(usize, (FieldElement, usize), usize, usize),
 }
 
 /// Compiles an ACIR circuit into an [R1CS] instance and associated [Solver].
@@ -234,12 +196,12 @@ impl R1CS {
 
     // Return the R1CS witness index corresponding to the AcirWitness provided, creating a new R1CS
     // witness (and builder) if required.
-    fn to_r1cs_witness(&mut self, acir_witness: AcirWitness) -> usize {
+    fn to_r1cs_witness_index(&mut self, acir_witness_index: AcirWitness) -> usize {
         self.acir_to_r1cs_witness_map
-            .get(&acir_witness.as_usize())
+            .get(&acir_witness_index.as_usize())
             .copied()
             .unwrap_or_else(|| {
-                self.add_witness(WitnessBuilder::Acir(acir_witness.as_usize()))
+                self.add_witness(WitnessBuilder::Acir(acir_witness_index.as_usize()))
             })
     }
 
@@ -303,8 +265,8 @@ impl R1CS {
                 .iter()
                 .take(expr.mul_terms.len() - 1)
                 .map(|(coeff, acir_witness0, acir_witness1)| {
-                    let witness0 = self.to_r1cs_witness(*acir_witness0);
-                    let witness1 = self.to_r1cs_witness(*acir_witness1);
+                    let witness0 = self.to_r1cs_witness_index(*acir_witness0);
+                    let witness1 = self.to_r1cs_witness_index(*acir_witness1);
                     (-*coeff, self.add_product(witness0, witness1))
                 })
                 .collect::<Vec<_>>();
@@ -313,16 +275,16 @@ impl R1CS {
             let (coeff, acir_witness0, acir_witness1) = &expr.mul_terms[expr.mul_terms.len() - 1];
             a = vec![(
                 FieldElement::from(*coeff),
-                self.to_r1cs_witness(*acir_witness0),
+                self.to_r1cs_witness_index(*acir_witness0),
             )];
-            b = vec![(FieldElement::one(), self.to_r1cs_witness(*acir_witness1))];
+            b = vec![(FieldElement::one(), self.to_r1cs_witness_index(*acir_witness1))];
         }
 
         // Extend with linear combinations
         linear.extend(
             expr.linear_combinations
                 .iter()
-                .map(|(coeff, acir_witness)| (coeff.neg(), self.to_r1cs_witness(*acir_witness))),
+                .map(|(coeff, acir_witness)| (coeff.neg(), self.to_r1cs_witness_index(*acir_witness))),
         );
 
         // Add constant by multipliying with constant value one.
@@ -357,263 +319,6 @@ impl R1CS {
     }
 }
 
-/// Mock transcript for testing purposes.
-pub struct MockTranscript {
-    count: u32,
-}
-
-impl MockTranscript {
-    pub fn new() -> Self {
-        Self {
-            count: 0,
-        }
-    }
-
-    pub fn append(&mut self, _value: FieldElement) {
-        self.count += 1;
-    }
-
-    pub fn draw_challenge(&mut self) -> FieldElement {
-        self.count +=1;
-        self.count.into()
-    }
-}
-
-pub struct R1CSSolver {
-    /// Indicates how to solve for each R1CS witness
-    pub witness_builders: Vec<WitnessBuilder>,
-
-    /// The length of each memory block
-    pub memory_lengths: BTreeMap<usize, usize>,
-}
-
-impl R1CSSolver {
-    pub fn new() -> Self {
-        Self {
-            witness_builders: vec![WitnessBuilder::Constant(FieldElement::one())],
-            memory_lengths: BTreeMap::new(),
-        }
-    }
-
-    /// Add a new witness to the R1CS solver.
-    pub fn add_witness_builder(&mut self, witness_builder: WitnessBuilder) {
-        self.witness_builders.push(witness_builder);
-    }
-
-    pub fn solve(&self, transcript: &mut MockTranscript, acir_witnesses: &WitnessMap<FieldElement>) -> Vec<FieldElement> {
-        let mut witness: Vec<Option<FieldElement>> = vec![None; self.num_witnesses()];
-        // The memory read counts for each block of memory
-        let mut memory_read_counts: BTreeMap<usize, Vec<u32>> = self.memory_lengths.iter().map(|(block_id, len)| (*block_id, vec![0u32; *len])).collect();
-        self.witness_builders.iter().enumerate().for_each(|(witness_idx, witness_builder)| {
-            assert_eq!(witness[witness_idx], None, "Witness {witness_idx} already set.");
-            let value = match witness_builder {
-                WitnessBuilder::Constant(c) => *c,
-                WitnessBuilder::Acir(acir_witness_idx) => {
-                    acir_witnesses[&AcirWitness(*acir_witness_idx as u32)]
-                },
-                WitnessBuilder::StaticMemoryRead(block_id, static_addr, value_acir_witness_idx) => {
-                    memory_read_counts.get_mut(&block_id).unwrap()[*static_addr] += 1;
-                    acir_witnesses[&AcirWitness(*value_acir_witness_idx as u32)]
-                },
-                WitnessBuilder::DynamicMemoryRead(block_id, addr_witness_idx, value_acir_witness_idx) => {
-                    let addr = witness[*addr_witness_idx].unwrap().try_to_u64().unwrap() as usize;
-                    memory_read_counts.get_mut(&block_id).unwrap()[addr] += 1;
-                    acir_witnesses[&AcirWitness(*value_acir_witness_idx as u32)]
-                },
-                WitnessBuilder::Challenge => {
-                    transcript.draw_challenge()
-                },
-                WitnessBuilder::Inverse(operand_idx) => {
-                    let operand: FieldElement = witness[*operand_idx].unwrap();
-                    operand.inverse()
-                },
-                WitnessBuilder::Product(operand_idx_a, operand_idx_b) => {
-                    let a: FieldElement = witness[*operand_idx_a].unwrap();
-                    let b: FieldElement = witness[*operand_idx_b].unwrap();
-                    a * b
-                },
-                WitnessBuilder::Sum(operands) => {
-                    operands.iter().map(|idx| witness[*idx].unwrap()).fold(FieldElement::zero(), |acc, x| acc + x)
-                },
-                WitnessBuilder::MemoryAccessCount(block_id, addr) => {
-                    let count = memory_read_counts.get(&block_id).unwrap()[*addr];
-                    FieldElement::from(count)
-                },
-                WitnessBuilder::LogUpDenominator(sz_challenge, (index_coeff, index), rs_challenge, value) => {
-                    let index = witness[*index].unwrap();
-                    let value = witness[*value].unwrap();
-                    let rs_challenge = witness[*rs_challenge].unwrap();
-                    let sz_challenge = witness[*sz_challenge].unwrap();
-                    let denominator = sz_challenge - (*index_coeff * index + rs_challenge * value);
-                    denominator
-                },
-            };
-            witness[witness_idx] = Some(value);
-            transcript.append(value);
-        });
-
-        // Complete witness with entropy.
-        // TODO: Use better entropy source and proper sampling.
-        // FIXME is this the desired behaviour?  Would an error be more appropriate if the solver fails to determine the witness?
-        let mut rng = rand::thread_rng();
-        let witness = witness
-            .iter()
-            .map(|f| f.unwrap_or_else(|| FieldElement::from(rng.gen::<u128>())))
-            .collect::<Vec<_>>();
-        witness
-    }
-
-    /// The number of witnesses in the R1CS instance.
-    /// This includes the constant one witness.
-    pub fn num_witnesses(&self) -> usize {
-        self.witness_builders.len()
-    }
-
-    /// Index of the constant 1 witness
-    pub fn witness_one(&self) -> usize {
-        0
-    }
-}
-
-/// Represents a R1CS constraint system.
-#[derive(Debug, Clone)]
-pub struct R1CSMatrices {
-    pub a: SparseMatrix<FieldElement>,
-    pub b: SparseMatrix<FieldElement>,
-    pub c: SparseMatrix<FieldElement>,
-}
-
-#[derive(Serialize, Deserialize)]
-struct MatrixEntry {
-    constraint: usize,
-    signal:     usize,
-    value:      String,
-}
-
-impl R1CSMatrices {
-    pub fn new() -> Self {
-        Self {
-            a: SparseMatrix::new(0, 1, FieldElement::zero()),
-            b: SparseMatrix::new(0, 1, FieldElement::zero()),
-            c: SparseMatrix::new(0, 1, FieldElement::zero()),
-        }
-    }
-
-    pub fn to_json(
-        &self,
-        num_public: usize,
-        witness: &[FieldElement],
-    ) -> Result<String, serde_json::Error> {
-
-        // Convert witness to string format
-        let witnesses = vec![witness
-            .iter()
-            .map(|w| w.to_string())
-            .collect::<Vec<String>>()];
-
-        let json_r1cs = JsonR1CS {
-            num_public,
-            num_variables: self.num_witnesses(),
-            num_constraints: self.num_constraints(),
-            a: Self::matrix_to_entries(&self.a),
-            b: Self::matrix_to_entries(&self.b),
-            c: Self::matrix_to_entries(&self.c),
-            witnesses,
-        };
-
-        serde_json::to_string_pretty(&json_r1cs)
-    }
-
-    fn matrix_to_entries(matrix: &SparseMatrix<FieldElement>) -> Vec<MatrixEntry> {
-        matrix.entries.iter().filter_map(|((row, col), value)| {
-            if !value.is_zero() {
-                Some(MatrixEntry {
-                    constraint: *row,
-                    signal:     *col,
-                    value:      value.to_string(),
-                });
-            }
-            None
-        }).collect()
-    }
-
-    pub fn write_json_to_file(
-        &self,
-        num_public: usize,
-        witness: &[FieldElement],
-        path: &str,
-    ) -> std::io::Result<()> {
-        let json = self
-            .to_json(num_public, witness)
-            .map_err(|e| std::io::Error::new(std::io::ErrorKind::Other, e.to_string()))?;
-
-        let mut file = File::create(path)?;
-        file.write_all(json.as_bytes())?;
-        Ok(())
-    }
-
-    /// The number of constraints in the R1CS instance.
-    pub fn num_constraints(&self) -> usize {
-        self.a.rows
-    }
-
-    /// The number of witnesses in the R1CS instance (including the constant one witness).
-    pub fn num_witnesses(&self) -> usize {
-        self.a.cols
-    }
-
-    /// Add a new witness to the R1CS instance, returning its index.
-    pub fn add_witness(&mut self) -> usize {
-        let next_witness_idx = self.num_witnesses();
-        self.grow_matrices(self.num_constraints(), self.num_witnesses() + 1);
-        next_witness_idx
-    }
-
-    // Increase the size of the R1CS matrices to the specified dimensions.
-    fn grow_matrices(&mut self, num_rows: usize, num_cols: usize) {
-        self.a.grow(num_rows, num_cols);
-        self.b.grow(num_rows, num_cols);
-        self.c.grow(num_rows, num_cols);
-    }
-
-    // Adds a new R1CS constraint.
-    pub fn add_constraint(
-        &mut self,
-        az: &[(FieldElement, usize)],
-        bz: &[(FieldElement, usize)],
-        cz: &[(FieldElement, usize)],
-    ) {
-
-        let next_constraint_idx = self.num_constraints();
-        let num_cols = self.num_witnesses();
-        self.grow_matrices(self.num_constraints() + 1, num_cols);
-
-        for (coeff, witness_idx) in az.iter().copied() {
-            self.a.set(next_constraint_idx, witness_idx, coeff)
-        }
-        for (coeff, witness_idx) in bz.iter().copied() {
-            self.b.set(next_constraint_idx, witness_idx, coeff)
-        }
-        for (coeff, witness_idx) in cz.iter().copied() {
-            self.c.set(next_constraint_idx, witness_idx, coeff)
-        }
-    }
-
-    /// Returns None if this R1CS instance is satisfied, otherwise returns the index of the first
-    /// constraint that is not satisfied.
-    pub fn test_satisfaction(&self, witness: &[FieldElement]) -> Option<usize> {
-        let az = mat_mul(&self.a, witness);
-        let bz = mat_mul(&self.b, witness);
-        let cz = mat_mul(&self.c, witness);
-        for (row, ((a_val, b_val), c_val)) in az.into_iter().zip(bz.into_iter()).zip(cz.into_iter()).enumerate() {
-            if a_val * b_val != c_val {
-                return Some(row);
-            }
-        }
-        None
-    }
-}
-
 /// Print the R1CS matrices and the ACIR -> R1CS witness map, useful for debugging.
 impl std::fmt::Display for R1CS {
     fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
@@ -629,21 +334,5 @@ impl std::fmt::Display for R1CS {
             writeln!(f, "")?;
         }
         writeln!(f, "{}", self.matrices)
-    }
-}
-
-/// Print the R1CS matrices and the ACIR -> R1CS witness map, useful for debugging.
-impl std::fmt::Display for R1CSMatrices {
-    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
-        if std::cmp::max(self.num_constraints(), self.num_witnesses()) > 15 {
-            println!("R1CS matrices too large to print");
-            return Ok(());
-        }
-        writeln!(f, "Matrix A:")?;
-        write!(f, "{}", self.a)?;
-        writeln!(f, "Matrix B:")?;
-        write!(f, "{}", self.b)?;
-        writeln!(f, "Matrix C:")?;
-        write!(f, "{}", self.c)
     }
 }
