+++ noir-r1cs/src/sparse_matrix.rs
@@ -1,5 +1,5 @@
 use std::{
-    collections::BTreeMap, fmt::{Debug, Display, Formatter}, ops::{Add, AddAssign, Index, IndexMut, Mul}
+    collections::BTreeMap, fmt::{Debug, Display, Formatter}, ops::{AddAssign, Index, IndexMut, Mul}
 };
 
 use acir::{AcirField, FieldElement};
@@ -142,39 +142,6 @@ where
         }
         result
     }
-}
-
-// Sparse dot product. `a` is assumed zero. `b` is assumed missing.
-pub fn sparse_dot<'a>(
-    a: impl Iterator<Item = (usize, &'a FieldElement)>,
-    b: &[Option<FieldElement>],
-) -> Option<FieldElement> {
-    let mut accumulator = FieldElement::zero();
-    for (col, &a) in a {
-        accumulator += a * b[col]?;
-    }
-    Some(accumulator)
-}
-
-// Returns a pair (i, f) such that, setting `b[i] = f`,
-// ensures `sparse_dot(a, b) = r`.
-fn solve_dot<'a>(
-    a: impl Iterator<Item = (usize, &'a FieldElement)>,
-    b: &[Option<FieldElement>],
-    r: FieldElement,
-) -> Option<(usize, FieldElement)> {
-    let mut accumulator = -r;
-    let mut missing = None;
-    for (col, &a) in a {
-        if let Some(b) = b[col] {
-            accumulator += a * b;
-        } else if missing.is_none() {
-            missing = Some((col, a));
-        } else {
-            return None;
-        }
-    }
-    missing.map(|(col, coeff)| (col, -accumulator / coeff))
 }
 
 pub fn mat_mul(a: &SparseMatrix<FieldElement>, b: &[FieldElement]) -> Vec<FieldElement> {
