+++ noir-r1cs/src/utils/mod.rs
@@ -1,3 +1,3 @@
 pub mod file_io;
 mod print_abi;
-pub use self::print_abi::{PrintAbi, PrintType};
+pub use self::print_abi::PrintAbi;
